S7comm-Project/s7comm-python/pyproject.toml
[project]
name = "s7comm-mcp"
version = "0.1.0"
description = "MCP server and tooling for Siemens S7 PLCs over the S7comm protocol."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "IndustriConnect" }
]
dependencies = [
    "anyio>=4.0.0",
    "mcp[cli]>=1.6.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.1.0",
    "python-snap7>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "ruff>=0.6.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
s7comm-mcp = "s7comm_mcp.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
